[societegenerale/sgpe] Go to the accounts page after execute_transfer on pro browser
When we reload the confirm_transfer page it returns us an error because
we try to reload the url with GET method (it is called with the POST
method in the module). To avoid that we simply go on the accounts
page at the end of the execute_transfer to reload a good page.

Closes: 60183@sibi 60200@sibi