[cmso] Fix transactions
The previous commit forced the content-type of every request to
application/json which broke the request to get transactions. With the
wrong header, the response defaults to the five last weeks of history
instead of the requested month. This resulted in duplicated transactions
as we got the same transactions for each request.

Closes: 16691@zendesk, 17238@zendesk, 16660@zendesk, 16684@zendesk,
17235@zendesk, 17240@zendesk, 16560@zendesk, 17275@zendesk, 16778@zendesk,
16648@zendesk, 17286@zendesk, 16569@zendesk, 16703@zendesk, 16572@zendesk,
17344@zendesk, 17416@zendesk, 17425@zendesk, 16688@zendesk, 16693@zendesk,
17470@zendesk, 16775@zendesk, 17501@zendesk, 17507@zendesk, 17505@zendesk,
17504@zendesk, 17502@zendesk, 17499@zendesk, 17519@zendesk, 17556@zendesk,
17574@zendesk, 17581@zendesk