[bp] Fixed parsing of some internal recipients
Some internal recipients cannot be parsed on the website and so, do not
appear in the iter_accounts. Those accounts do not have the id in their
label, so we need to change where the id is gathered from.
The `value` of each `option` of the select is the same as the id we
can parse from the label, so we can directly get the id from the `value`
attribute.

Closes: 53010@sibi 53027@sibi