[s2e] use "show all" button for documents + improvements
There might be more documents than displayed on the first documents
page. A button "Afficher tout" exists which shows everything. It might
be missing if we clicked it already (warning: the site is stateful for
the session).

Also, the id from URL might not reflect the document type, it has been
observed to be "EDI/..." for various document types, without "RDC" or
anything meaningful. So, type using the label instead.