[caissedepargne] do not compare bytes and unicode on cookies
In python2, cookies from a cookiejar are bytes, but we compare them to a
unicode string (unicode_literals). This is not a problem until there are
non-ascii bytes.

Unfortunately, sometimes the caissedepargne/palatine site returns
cookies in utf-8. So we end with non-ascii bytes.

In python3, it doesn't crash, but requests or the lower layers interpret
cookies as latin-1 instead of utf-8, which yields a few garbled
characters, but fortunately we do not care about that part of the
cookie.

Closes: 57830@sibi